fix: remove macro crate in favor of singular macro crate
