--- docker-images/syntax-highlighter/crates/sg-macros/BUILD.bazel
@@ -1,30 +1,0 @@
-load("@crate_index//:defs.bzl", "aliases", "all_crate_deps")
-load("@rules_rust//rust:defs.bzl", "rust_proc_macro", "rust_test")
-
-rust_proc_macro(
-    name = "sg-macros",
-    srcs = ["src/lib.rs"],
-    aliases = aliases(),
-    proc_macro_deps = all_crate_deps(
-        proc_macro = True,
-    ),
-    visibility = ["//docker-images/syntax-highlighter:__subpackages__"],
-    deps = all_crate_deps(
-        normal = True,
-    ),
-)
-
-rust_test(
-    name = "unit_test",
-    aliases = aliases(
-        normal_dev = True,
-        proc_macro_dev = True,
-    ),
-    crate = ":sg-macros",
-    proc_macro_deps = all_crate_deps(
-        proc_macro_dev = True,
-    ),
-    deps = all_crate_deps(
-        normal_dev = True,
-    ),
-)
