--- docker-images/syntax-highlighter/crates/sg-macros/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "sg-macros"
-version = "0.1.0"
-edition = "2021"
-autotests = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1"
-syn = "1"
-
-[dev-dependencies]
-pretty_assertions = "1"
